+++ ios/Classes/FlutteryAudioPlugin.h
@@ -1,4 +1,5 @@
 #import <Flutter/Flutter.h>
+#import "AudioPlayer.h"
 
-@interface FlutteryAudioPlugin : NSObject<FlutterPlugin>
+@interface FlutteryAudioPlugin : NSObject<FlutterPlugin, AudioPlayerListener>
 @end