+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## [0.0.1] - TODO: Add release date.
 
-* TODO: Describe initial release.
+Initial Release.  Audio can be streamed on Android and iOS.
+
+* Flutter Widgets created: Audio, AudioPlaylist.
+* Imperative Object available: AudioPlayer.
+* URL playback is supported.